added the backend code for CRUD operation of backlogs, project tasks

+++ project-management-tool/src/main/java/com/simpleUI/projectmanagementtool/repository/ProjectRepository.java
@@ -6,9 +6,11 @@ import org.springframework.stereotype.Repository;
 import com.simpleUI.projectmanagementtool.domain.Project;
 
 @Repository
-public interface ProjectRepository extends CrudRepository<Project, Integer> {
+public interface ProjectRepository extends CrudRepository<Project, Long> {
 
 	Project findByprojectIdentifier(String projectIdentifier);
+	
+
 
 }
